move KillHandler import to util module
