+++ pyabc/sampler/redis_eps/work.py
@@ -8,7 +8,6 @@ import cloudpickle as pickle
 from redis import StrictRedis
 
 from ..util import any_particle_preliminary
-from .cli import KillHandler
 from .cmd import (
     ALL_ACCEPTED,
     ANALYSIS_ID,
@@ -27,6 +26,7 @@ from .cmd import (
     SSA,
     idfy,
 )
+from .util import KillHandler
 
 logger = logging.getLogger('ABC.Sampler')
